Update README.md
Hi! 👋 I'm Nishtha, a B.Tech 2nd-year student with a passion for coding and development. 🚀

I created Girdhar Extension as a powerful multi-language execution tool for VS Code, making it easier for developers to run different programming languages within a single environment. 🖥️✨

The name "Girdhar" holds a special place in my heart 💙—I named it after my dear friend, as a token of appreciation for their constant support and motivation. This project is a reflection of my learning journey in tech, and I hope it helps other developers in their workflow!

Happy Coding! 😊💻